requirements.txt
discord.py==1.7.3
mss==6.1.0
opencv-python==4.5.4.60
pynput==1.7.4
requests==2.26.0
python-dotenv==0.19.1
Pillow==9.0.0
pystray==0.18.0
plyer==2.0.0
pytube==11.0.2
pycaw==20181226
speedtest-cli==2.1.3
pyperclip==1.8.2
win10toast==0.9
